+++ NEWS.md
@@ -126,6 +126,8 @@
 * The {lintr} configuration file is now searched in the system's user configuration path; the lintr config filename can
   also be configured explicitly by setting the environment variable `R_LINTR_LINTER_FILE` (#460, @klmr)
 
+* Errors in the {lintr} configuration file now produce more informative error messages (#886, @AshesITR)
+
 ### New linters
 
 * `matrix_apply_linter()` recommends use of dedicated `rowSums()`, `colSums()`, `colMeans()`, `rowMeans()` over `apply(., MARGIN, sum)` or `apply(., MARGIN, mean)`. The recommended alternative is much more efficient and more readable (#1869, @Bisaloo).